Revert "Greg/scripts"
