--- .babel-register
@@ -1,15 +1,0 @@
-/*
-  See
-    https://github.com/babel/babel/issues/8652
-    https://github.com/babel/babel/pull/6027
-  Babel isn't currently configured by default to read .ts files and
-  can only be configured to do so via cli or configuration below.
-  
-  This file is used by mocha to interpret test files using a properly
-  configured babel.
-
-  This can (probably) be removed in babel 8.x.
-*/
-require('@babel/register')({
-  extensions: ['.ts'],
-})
