--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/env",
-      {"targets": {"node":"current"}}
-    ],
-    "@babel/typescript"
-  ],
-  "plugins": [
-    "@babel/proposal-class-properties",
-    "@babel/proposal-object-rest-spread",
-    "@babel/plugin-transform-runtime"
-  ]
-}
