+++ .gitignore
@@ -1,3 +1,1 @@
 node_modules
-yarn-error.log
-lib
