+++ index.ts
@@ -1,0 +1,6 @@
+// Relative imports
+import Main from './lib';
+
+Main()
+  .then(() => { console.log("attalus executed without errors!")})
+  .catch((err: any) => { console.log("attalus executed with errors: ", err)});
