+++ lib/attalus.ts
@@ -1,7 +1,8 @@
-import ethers, {Wallet} from "ethers";
-import {JsonRpcProvider} from "ethers/providers";
+import {Wallet} from "ethers";
 
-import config from "./config";
+const ethers = require("ethers");
+import {JsonRpcProvider} from "ethers/providers";
+import config from "../config"
 import {bn, parseGwei} from "./utilities/conversion";
 
 const connect = (url: string): JsonRpcProvider => {
@@ -13,11 +14,11 @@ const connect = (url: string): JsonRpcProvider => {
 */
 const generateWallets = async (num: number) => {
   const wallets = [];
-  for (let i = 0; i < num; i++) {
-    const wallet: Wallet = ethers.Wallet.createRandom();
+  for (let i: number = 0; i < num; i++) {
+    const wallet = new ethers.Wallet.createRandom();
     wallets.push(wallet);
     const address = await wallet.getAddress();
-    console.log(`Created wallet with address ${address}`);
+    console.log(`Created wallet with address ${address}`)
     
   }
   console.log('\n');
@@ -28,7 +29,7 @@ const generateWallets = async (num: number) => {
 /*
 * Funds wallets in array wallets with mainWallet
 */
-const fundWallets = async (wallets: any[], mainWallet: any): Promise<string[]> => {
+const fundWallets = async (wallets: Array<any>, mainWallet: any): Promise<string[]> => {
   //send each wa]llet the max possible gas amount for each transaction + the amount of the transaction as specified in config
   const numTransactions = bn(config.numTransactions/config.numWallets);
   const transactionAmount = bn(config.amount).mul(numTransactions);
@@ -38,7 +39,7 @@ const fundWallets = async (wallets: any[], mainWallet: any): Promise<string[]> =
   const txHashes: string[] = [];
   const numWallets = wallets.length;
   let nonce = await mainWallet.getTransactionCount();
-  for (let i = 0; i < numWallets; i++) {
+  for (let i: number = 0; i < numWallets; i++) {
     const dest = await wallets[i].getAddress();
     const tx = {
       nonce: nonce,
@@ -49,7 +50,7 @@ const fundWallets = async (wallets: any[], mainWallet: any): Promise<string[]> =
       chainId: config.chainId
     };
     const txResponse = await mainWallet.sendTransaction(tx);
-    console.log(`sent transaction to fund address ${dest}`);
+    console.log(`sent transaction to fund address ${dest}`)
     txHashes.push(txResponse.hash);
     nonce += 1;
   }
@@ -57,19 +58,19 @@ const fundWallets = async (wallets: any[], mainWallet: any): Promise<string[]> =
 
   return txHashes;
 
-};
+}
 
 /*
 * Creates and broadcasts batches of transactions from wallets in array wallets to provider
 */
-const batchTxs = async (wallets: any[], provider: JsonRpcProvider) => {
+const batchTxs = async (wallets: Array<any>, provider: JsonRpcProvider) => {
   //we want to split the transactions equally among the wallets to be sent from.
   const numTransactions = Math.ceil(config.numTransactions/config.numWallets);
   const amount = bn(config.amount);
   const txs: any = [];
   const numWallets = wallets.length; 
-  console.log("Broadcasting transactions..."); 
-  for (let i = 0; i < wallets.length; i++) {
+  console.log("Broadcasting transactions...") 
+  for (let i: number = 0; i < wallets.length; i++) {
     const sender: Wallet = new ethers.Wallet(wallets[i].privateKey, provider);
 
     let nonce = 0;
@@ -89,7 +90,7 @@ const batchTxs = async (wallets: any[], provider: JsonRpcProvider) => {
       sender.sendTransaction(tx);
       txs.push(tx); 
     }
-  }
+    }
   console.log(`\nCreated and broadcasted ${txs.length} transactions.`);
   return txs;
 };
@@ -97,30 +98,34 @@ const batchTxs = async (wallets: any[], provider: JsonRpcProvider) => {
 /*
 * Broadcasts transactions from mainWallet to provider to call testOpcodes() at all known deployed contract addresses
 */
-const testOpcodes = async (provider: JsonRpcProvider, contractAddresses: any[], mainWallet) => {
+const testOpcodes= async (provider: JsonRpcProvider, contractAddresses: Array<any>, mainWallet) => {
 
   let nonce = await mainWallet.getTransactionCount();
   let txResponses = [];
   console.log("calling testOpcodes...");
 	
-  for (let i = 0; i < contractAddresses.length; i++){
-    const tx = {
-      nonce: nonce,
-      to: contractAddresses[i],
-      value: 0,
-      gasLimit: bn(config.maxGas),
-      gasPrice: parseGwei(config.gasPrice),
-      chainId: config.chainId,
-      //ABI for all contracts is the same, testOpcodes is 0x391521f4
-      data: "0x391521f4"
-    };
-    const txResponse = await mainWallet.sendTransaction(tx);
-    txResponses.push(txResponse.hash);
-    nonce += 1;
-  }
+  for (let i: number = 0; i < contractAddresses.length; i++){
+        const tx = {
+            nonce: nonce,
+            to: contractAddresses[i],
+            value: 0,
+            gasLimit: bn(config.maxGas),
+            gasPrice: parseGwei(config.gasPrice),
+            chainId: config.chainId,
+            //ABI for all contracts is the same, testOpcodes is 0x391521f4
+            data: "0x391521f4"
+        };
+        const txResponse = await mainWallet.sendTransaction(tx);
+        txResponses.push(txResponse.hash)
+        nonce += 1;
+    }
 
-  return txResponses;
-};
+    return txResponses;
+
+
+  
+
+}
 
 export {
   connect,
@@ -128,4 +133,4 @@ export {
   fundWallets,
   batchTxs,
   testOpcodes
-};
+}
