+++ package.json
@@ -4,37 +4,20 @@
   "description": "",
   "main": "index.ts",
   "scripts": {
-    "prebuild": "rm -rf lib",
-    "build": "babel src -x .ts -d lib --source-maps",
-    "check-types": "tsc --noEmit --incremental",
-    "lint": "eslint --ext .ts src",
-    "lint-fix": "eslint --ext .ts src --fix",
-    "start": "yarn build && node ./lib/index.js"
+    "start": "node_modules/.bin/ts-node index.ts"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/cli": "^7.5.0",
-    "@babel/core": "^7.5.4",
-    "@babel/plugin-proposal-class-properties": "^7.5.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.5.4",
-    "@babel/plugin-transform-runtime": "^7.5.0",
-    "@babel/preset-env": "^7.5.4",
-    "@babel/preset-typescript": "^7.3.3",
-    "@babel/runtime": "^7.5.4",
     "@types/chai": "^4.1.4",
     "@types/mocha": "^5.2.5",
-    "@typescript-eslint/eslint-plugin": "^1.11.0",
-    "@typescript-eslint/parser": "^1.11.0",
     "chai": "^4.1.2",
-    "eslint": "5.14.1",
     "mocha": "^5.2.0",
     "nyc": "^13.0.1",
     "ts-node": "^7.0.1",
     "typescript": "^3.0.3"
   },
   "dependencies": {
-    "ethers": "^4.0.0-beta.14",
-    "fs-extra": "^8.1.0"
+    "ethers": "^4.0.0-beta.14"
   }
 }
